Flash the status LED when the button is pressed.

+++ control.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"github.com/mrmorphic/hwio"
 	"log"
 	"os"
@@ -10,8 +11,16 @@ import (
 
 const (
 	switch_pin_string string        = "GPIO8"
+	out_pin_string    string        = "GPIO7"
 	sleep             time.Duration = 500 * time.Millisecond
 )
+
+var pretend bool
+
+func init() {
+	flag.BoolVar(&pretend, "pretend", false, "Don't really chut down, ust flash the LED on button press. Useful for debugging.")
+	flag.Parse()
+}
 
 func main() {
 	switch_pin, pin_err := hwio.GetPinWithMode(switch_pin_string, hwio.INPUT_PULLDOWN)
@@ -27,12 +36,33 @@ func main() {
 		}
 
 		if val == 1 {
+			done_chan := make(chan int)
+			go func() {
+				out_pin, err := hwio.GetPinWithMode(out_pin_string, hwio.OUTPUT)
+
+				if err != nil {
+					log.Fatal(err)
+				}
+
+				for i := 0; i < 10; i++ {
+					hwio.DigitalWrite(out_pin, hwio.LOW)
+					time.Sleep(150 * time.Millisecond)
+					hwio.DigitalWrite(out_pin, hwio.HIGH)
+					time.Sleep(150 * time.Millisecond)
+				}
+				done_chan <- 1
+			}()
+
 			log.Println("Shutting down the system as due to power button press")
-			cmd := exec.Command("systemctl", "poweroff")
-			err := cmd.Run()
-			if err != nil {
-				log.Fatal(err)
+			if !pretend {
+				cmd := exec.Command("systemctl", "poweroff")
+				err := cmd.Run()
+				if err != nil {
+					log.Fatal(err)
+				}
 			}
+
+			<-done_chan
 			os.Exit(0)
 		}
 		time.Sleep(sleep)
